ios adjustments
