+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.2.1",
     "react-scripts": "^5.0.1",
     "react-spinners": "^0.14.1",
     "three": "^0.167.1",
