+++ src/useSliderColors/style.css
@@ -14,16 +14,18 @@
   }
   
   .slider-item {
-    width: 50px;
-    height: 50px;
+    width: 30px;
+    height: 30px;
     cursor: pointer;
   }
+
   @media screen and (max-width: 768px) {
     .slider-item {
       width: 30px;
       height: 30px;
     }
   }
+  
   @media screen and (max-width: 380px) {
     .slider-item {
       width: 20px;
